Add files via upload
Varun is the founder of Event Management Company, "Sparsh Service".
In the company, all the financial transaction are carried out online and Varun has strictly
insisted his staff do any transactions on web browsers that supports AES encryption
nubbers. A number is an AES number if it has exactly four divisors. in other words,
there are exactly four numbers that divide into it evenly. For example, 10 is an ASE
number because it has exactly four divisors (1,2,5,10). 12 is not an AES number because
it has too many divisors (1,2,3,4,6,12). 11 is not an AES number either. write a program
that counts how many numbers from that range are AES numbers.

input range: 1,20
output : 5
input range: 50,1000
output : 277